refactor: split components to pass formatting checks
